requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.10
billiard==4.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
dill==0.4.0
Django==5.2.3
django-anymail==13.0
django-cors-headers==4.7.0
django-prometheus==2.4.1
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
idna==3.10
importlib_metadata==8.7.0
isort==6.0.1
kombu==5.5.4
mccabe==0.7.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-django==0.55b1
opentelemetry-instrumentation-wsgi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
packaging==25.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==5.29.5
psycopg2==2.9.10
PyJWT==2.9.0
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.23.0
